Refactor unit test namespaces.
