+++ tests/unit/lib/DataWarehouse/Export/CsvFileWriterTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace UnitTesting\DataWarehouse\Export;
+namespace UnitTests\DataWarehouse\Export;
 
 use CCR\Log;
 use DataWarehouse\Export\FileWriter\CsvFileWriter;
 use PHPUnit_Framework_TestCase;
 use Psr\Log\LoggerInterface;
-use TestHarness\TestFiles;
+use IntegrationTests\TestHarness\TestFiles;
 
 /**
  * Test data warehouse export CSV file writer.
@@ -33,7 +33,7 @@ class CsvFileWriterTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @return \TestHarness\TestFiles
+     * @return TestFiles
      */
     public function getTestFiles()
     {
