+++ tests/unit/lib/DataWarehouse/VisualizationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace UnitTesting\DataWarehouse;
+namespace UnitTests\DataWarehouse;
 
 class VisualizationTest extends \PHPUnit_Framework_TestCase
 {
