+++ tests/unit/lib/ETL/DataEndpoint/FileTest.php
@@ -7,11 +7,10 @@
  * ------------------------------------------------------------------------------------------
  */
 
-namespace UnitTesting\ETL\DataEndpoint;
+namespace UnitTests\ETL\DataEndpoint;
 
 use Exception;
 use CCR\Log;
-// use ETL\DataEndpoint\File;
 use ETL\DataEndpoint;
 use ETL\DataEndpoint\DataEndpointOptions;
 use Psr\Log\LoggerInterface;
