+++ tests/unit/lib/ETL/DbModel/TableTest.php
@@ -3,7 +3,7 @@
  * Test ETL table database models.
  */
 
-namespace UnitTesting\ETL\Configuration;
+namespace UnitTests\ETL\DbModel;
 
 use CCR\Log;
 use ETL\DbModel\Table;
