+++ tests/unit/lib/LogTest.php
@@ -7,7 +7,7 @@
  * ------------------------------------------------------------------------------------------
  */
 
-namespace UnitTesting;
+namespace UnitTests;
 
 use CCR\Log;
