+++ tests/unit/lib/Realm/RealmTest.php
@@ -3,7 +3,7 @@
  * Unit tests for Realm class.
  */
 
-namespace UnitTesting\Realm;
+namespace UnitTests\Realm;
 
 use CCR\Log as Logger;
 use Realm\Realm;
